Functions for Declared Missing Values

A set of functions to declare labels and missing values, coupled with associated functions to create (weighted) tables of frequencies and various other summary measures. Some of the base functions are rewritten to make use of the specific information about the missing values, most importantly to distinguish between empty and declared missing values. Many functions have a similar functionality with the corresponding ones from packages "haven" and "labelled". The aim is to ensure as much compatibility as possible with these packages, while offering an alternative in the objects of class "declared".

Details

Package:	declared

Type:	Package

Version:	0.27

Date:	2026-08-08

License:	GPL-v3

Author

Adrian Dusa

Maintainer: Adrian Dusa (dusa.adrian@unibuc.ro)